Extract InfoRequest::TitleValidation concern
Title validation has quite a few pieces, so extracting it out into a
concern helps to understand it as a whole.

I’ll also want to use this to validate DraftInfoRequest#title.

Extract specs to a somewhat namespaced example so that we can reuse
elsewhere where these validations matter.

Also does some minor cleanup of style.